+++ source/extensions/lightspeed.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.2]
+### Added
+- Added dictionary for Remix Categories
+
 ## [1.0.1]
 ### Changed
 - Changed repo link