+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.1]
+### Added
+- Categories icon
+
 ## [1.5.0] - 2024-04-23
 ### Added
 - Teleport icon